requirements.txt
wheel
Pillow
